Added Github Actions, moved to stacs-ci.
