+++ .github/workflows/check.yml
@@ -1,0 +1,30 @@
+name: Check
+
+on:
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  check:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Configure Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.9
+
+      - name: Install Tox
+        run: |
+          python -m pip install --upgrade pip wheel setuptools
+          pip install tox
+
+      - name: Run Linter
+        run: |
+          tox -e lint
+
+      - name: Run Tests
+        run: |
+          tox -e test
