+++ .github/workflows/publish.yml
@@ -1,0 +1,50 @@
+name: Publish
+
+on:
+  release:
+    types: [published]
+
+jobs:
+  publish:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Configure Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.9
+
+      - name: Extract version
+        run: |
+          pip install .
+          STACS_CI_VERSION="$(python -c 'from stacs.ci.__about__ import __version__ ; print(__version__, end="")')"
+          echo "IMAGE_VERSION=${STACS_CI_VERSION}" >> "${GITHUB_ENV}"
+
+      - name: Login to DockerHub
+        uses: docker/login-action@v1 
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+
+      - name: Build and push Docker image
+        id: docker_build
+        uses: docker/build-push-action@v2
+        with:
+          context: .
+          push: true
+          tags: stacscan/stacs-ci:latest,stacscan/stacs-ci:${{ env.IMAGE_VERSION }}
+          build-args: |
+            VERSION=${{ env.IMAGE_VERSION }}
+
+      - name: Build Python package
+        run: |
+          python -m pip install --upgrade pip wheel setuptools
+          pip install build
+          python -m build --sdist --outdir dist/ .
+
+      - name: Publish Python package
+        uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
+        with:
+          user: __token__
+          password: ${{ secrets.PYPI_TOKEN }}
