+++ LICENSE
@@ -1,6 +1,6 @@
 BSD 3-Clause License
 
-Copyright (c) 2021, Peter Adkins
+Copyright (c) 2022, Peter Adkins
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
