--- action.yml
@@ -1,11 +1,0 @@
-# action.yml
-name: 'STACS Scan'
-description: 'Scan a directory for static tokens and credentials.'
-branding:
-  icon: 'check-circle'
-  color: 'blue'
-
-inputs:
-  scan-directory:
-    description: 'The subdirectory to scan'
-    default: ''
