--- examples/stacs.ignore.json
@@ -1,9 +1,0 @@
-{
-    "include": [],
-    "ignore": [
-        {
-            "pattern": "test/fixtures/shadow",
-            "reason": "Test shadow file contain example credentials."
-        }
-    ]
-}
