+++ stacs/ci/__about__.py
@@ -1,10 +1,10 @@
-"""Static Token And Credential Scanner Integrations.
+"""Static Token And Credential Scanner CI integrations.
 
 SPDX-License-Identifier: BSD-3-Clause
 """
 
-__title__ = "stacs-integration"
-__summary__ = "Static Token And Credential Scanner Integrations."
+__title__ = "stacs-ci"
+__summary__ = "Static Token And Credential Scanner CI Integrations."
 __version__ = "0.1.0"
 __author__ = "Peter Adkins"
 __uri__ = "https://www.github.com/stacscan/stacs-integration/"
