+++ stacs/ci/__init__.py
@@ -1,0 +1,12 @@
+"""Static Token And Credential Scanner CI integrations.
+
+SPDX-License-Identifier: BSD-3-Clause
+"""
+
+from stacs.ci import __about__  # noqa:F401
+from stacs.ci import constants  # noqa:F401
+from stacs.ci import diff  # noqa:F401
+from stacs.ci import exceptions  # noqa:F401
+from stacs.ci import generic  # noqa:F401
+from stacs.ci import github  # noqa:F401
+from stacs.ci import models  # noqa:F401
