+++ stacs/ci/generic/__init__.py
@@ -1,0 +1,3 @@
+"""STACS Generic Integration."""
+
+from stacs.ci.generic import fail  # noqa: F401.
