+++ stacs/ci/github/__init__.py
@@ -1,0 +1,8 @@
+"""STACS Github integration.
+
+SPDX-License-Identifier: BSD-3-Clause
+"""
+
+from stacs.ci.github import client  # noqa: F401
+from stacs.ci.github import constants  # noqa: F401
+from stacs.ci.github import pr  # noqa: F401
