--- stacs/integration/__init__.py
@@ -1,12 +1,0 @@
-"""Static Token And Credential Scanner Integrations.
-
-SPDX-License-Identifier: BSD-3-Clause
-"""
-
-from stacs.integration import __about__  # noqa:F401
-from stacs.integration import constants  # noqa:F401
-from stacs.integration import diff  # noqa:F401
-from stacs.integration import exceptions  # noqa:F401
-from stacs.integration import generic  # noqa:F401
-from stacs.integration import github  # noqa:F401
-from stacs.integration import models  # noqa:F401
