--- stacs/integration/generic/__init__.py
@@ -1,3 +1,0 @@
-"""STACS Generic Integration."""
-
-from stacs.integration.generic import fail  # noqa: F401.
