--- stacs/integration/github/__init__.py
@@ -1,8 +1,0 @@
-"""STACS Github integrations.
-
-SPDX-License-Identifier: BSD-3-Clause
-"""
-
-from stacs.integration.github import client  # noqa: F401
-from stacs.integration.github import constants  # noqa: F401
-from stacs.integration.github import pr  # noqa: F401
