+++ tests/fixtures/ci/credential.txt
@@ -1,0 +1,6 @@
+It contains a bunch of things, including static credentials. See?
+
+aws_access_key_id=AKIAZJ3PDTI2R9JLXCTH
+aws_secret_key_id=39isgIVkaUNhaUFZz+U3ENRpXnAmfh/KbPlUIWiQ
+
+Told you!
