+++ tests/fixtures/ci/none.txt
@@ -1,0 +1,39 @@
+Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque vel sapien sed est
+mattis cursus. Quisque pellentesque ligula vel fringilla pellentesque. Nunc quis purus
+orci. Vivamus diam est, volutpat sed dictum a, finibus sit amet lorem. Nulla hendrerit
+suscipit ex, vitae pharetra lacus pulvinar non. Suspendisse potenti. Duis facilisis
+nulla vel felis egestas imperdiet. Integer in ligula sem. Proin ultricies sed ex sit
+amet euismod. Praesent placerat molestie lacus et consectetur.
+
+Duis ultrices ante ac gravida pretium. Donec felis lectus, fringilla ac lobortis at,
+elementum id leo. Donec rhoncus mattis eros, ac faucibus nulla mattis facilisis. Ut et 
+mi orci. Pellentesque ut scelerisque urna. In augue mi, semper in tellus vitae, porta
+iaculis turpis. Integer feugiat sapien non ex maximus, vel mattis augue faucibus.
+Pellentesque ut feugiat risus. Cras quam leo, cursus eget dui ut, luctus aliquam ligula.
+Mauris ac felis pharetra, ornare lacus sit amet, facilisis elit.
+
+Mauris sit amet nisi nec nunc molestie tempor at a turpis. Suspendisse et purus odio.
+Suspendisse eu justo ut tellus fermentum luctus eu in nunc. Cras rhoncus rutrum
+ullamcorper. Nam tempor dui id laoreet consectetur. Maecenas tincidunt, elit et iaculis
+varius, quam nisl lacinia nibh, ut aliquam sem eros ac est. Cras et dictum dolor, semper
+tristique neque. Cras sed eleifend diam. Praesent malesuada dictum eleifend. Quisque nec
+pharetra ex. Aenean in nunc ullamcorper, cursus libero pulvinar, hendrerit nulla.
+
+Mauris accumsan erat quis augue faucibus, ac rutrum odio tincidunt. Quisque porta ipsum
+nec nunc porta, nec vestibulum massa imperdiet. Interdum et malesuada fames ac ante
+ipsum primis in faucibus. Interdum et malesuada fames ac ante ipsum primis in faucibus.
+Phasellus eu sapien tortor. Etiam rhoncus augue sapien, vitae posuere diam rhoncus ac.
+Sed eleifend sed ex imperdiet rhoncus. Class aptent taciti sociosqu ad litora torquent
+per conubia nostra, per inceptos himenaeos. Cras faucibus hendrerit ex eu tincidunt.
+Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos
+himenaeos. Cras gravida congue quam, eget hendrerit est bibendum eget. Class aptent
+taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Vivamus
+eget vehicula purus, a rutrum ipsum.
+
+Vestibulum blandit volutpat neque. Suspendisse semper lectus odio, eu ornare nisi mattis
+id. Pellentesque dignissim scelerisque urna, sit amet dignissim nibh consequat id. Proin
+et sodales arcu. Nullam nec viverra metus. Mauris at ligula arcu. In pellentesque
+blandit libero, a feugiat justo sodales eu. Donec vulputate dolor a sem sagittis, at
+congue mi feugiat. Nam maximus, felis quis pellentesque dignissim, eros leo accumsan
+orci, in accumsan tellus libero lacinia elit. Sed quis egestas nibh. Suspendisse
+posuere lobortis bibendum. Nam a lobortis nunc.
